campground authorizations

+++ routes/campgrounds.js
@@ -30,13 +30,14 @@ router.get('/new', isLoggedIn,(req, res) => {
 router.post('/', validateCampground, isLoggedIn, catchAsync(async (req, res, next) => {
     // if (!req.body.campground) throw new ExpressError('Invalid Campground Data', 400);
     const campground = new Campground(req.body.campground);
+    campground.author = req.user._id
     await campground.save();
     // req.flash('success', 'Successfully made a new campground!');
     res.redirect(`/campgrounds/${campground._id}`)
 }))
 
 router.get('/:id', isLoggedIn, catchAsync(async (req, res,) => {
-    const campground = await Campground.findById(req.params.id).populate('reviews');
+    const campground = await Campground.findById(req.params.id).populate('reviews').populate('author');
     if (!campground) {
         // req.flash('error', 'Cannot find that campground!');
         return res.redirect('/campgrounds');
@@ -45,7 +46,11 @@ router.get('/:id', isLoggedIn, catchAsync(async (req, res,) => {
 }));
 
 router.get('/:id/edit', isLoggedIn,catchAsync(async (req, res) => {
-    const campground = await Campground.findById(req.params.id)
+    const { id } = req.params;
+    const campground = await Campground.findById(id)
+    if(!campground.author.equals(req.user._id)) {
+        return res.redirect(`/campgrounds/${id}`)
+    }
     if (!campground) {
         // req.flash('error', 'Cannot find that campground!');
         return res.redirect('/campgrounds');
@@ -53,15 +58,26 @@ router.get('/:id/edit', isLoggedIn,catchAsync(async (req, res) => {
     res.render('campgrounds/edit', { campground });
 }))
 
+//edit campground
 router.put('/:id', validateCampground,isLoggedIn, catchAsync(async (req, res) => {
+    
     const { id } = req.params;
-    const campground = await Campground.findByIdAndUpdate(id, { ...req.body.campground });
+    const campground = await Campground.findById(id)
+    if(!campground.author.equals(req.user._id)) {
+        return res.redirect(`/campgrounds/${id}`)
+    }
+    const camp = await Campground.findByIdAndUpdate(id, { ...req.body.campground });
     // req.flash('success', 'Successfully updated campground!');
     res.redirect(`/campgrounds/${campground._id}`)
 }));
 
+//delete campground
 router.delete('/:id', isLoggedIn, catchAsync(async (req, res) => {
     const { id } = req.params;
+    const campground = await Campground.findById(id)
+    if(!campground.author.equals(req.user._id)) {
+        return res.redirect(`/campgrounds/${id}`)
+    }
     await Campground.findByIdAndDelete(id);
     // req.flash('success', 'Successfully deleted campground')
     res.redirect('/campgrounds');
